consolidated IncomeAdder and ExpenseAdder components into one - ItemAdder component
