+++ client/src/components/index.tsx
@@ -2,7 +2,6 @@ export { default as Navigation } from "./Navigation/Navigation";
 export { default as BudgetNavbar } from "./BudgetNavbar/BudgetNavbar";
 export { default as IncomeTracker } from "./IncomeTracker/IncomeTracker";
 export { default as IncomeItem } from "./IncomeItem/IncomeItem";
-export { default as IncomeAdder } from "./IncomeAdder/IncomeAdder";
 export { default as ExpenseItem } from "./ExpenseItem/ExpenseItem";
 export { default as ExpenseTracker } from "./ExpenseTracker/ExpenseTracker";
-export { default as ExpenseAdder } from "./ExpenseAdder/ExpenseAdder";
+export { default as ItemAdder } from "./ItemAdder/ItemAdder";
